refactor: move `hash` to `fingerprint` and `hashbytes` to `hash`
These operations were confusingly named. This commit changes the
existing `hash` API and corresponding `Hash` operation to be
`fingerprint` and `Fingerprint`, respectively.

`HashBytes` was also confusing, partially because the `Bytes` refers to
the output type instead of the input type. That is split into two
operations `HashBytes` and `HashString`, the suffix of which refers to
the *input* type. These operations are exposed under a single `.hash()`
API on `binary` and `string` value expressions.

Finally, because there are a large variety of ways to hash and backends
differ enormously on what algorithms they provide and how those
algorithsm are spelled, the `how` parameter to the new `hash` API is now
an arbitrary string expression.

BREAKING CHANGE: Change existing `.hash()` calls to `.fingerprint()` and existing `.hashbytes(how=...)` calls to `.hash(how=...)`.