refactor(polars): update the polars backend to use the new relational abstractions (#7868)
Bare minimum changes required for the polars backend to work with the
new relational operations. Since polars' join API follows the same
semantics as pandas, I'm using the pandas specific rewrites here.

In the future we may want to rewrite the compiler similarly to the
pandas one: using `node.map()` and `Dispatched`.