test(markers): add tests for custom markers
I had a thought about a potential edge case with `None` being the only
listed "Exception" in a `raises` argument.  This fixes the edge case and
also adds some simple tests to make sure our custom markers correctly
handle when an imported backend exception is set to `None` due to a
missing import.